encapsulate some logic and ui in LoadedLogsUi
