+++ src/plot/plot_ui.rs
@@ -1,13 +1,14 @@
-use chrono::NaiveDateTime;
-use egui::{Color32, Key, RichText, TextEdit};
+use egui::{Color32, RichText};
 use egui_phosphor::regular;
+use loaded_logs::LoadedLogsUi;
 
 use crate::app::PlayBackButtonEvent;
 
 use super::{
     axis_config::AxisConfig, play_state::PlayState, plot_visibility_config::PlotVisibilityConfig,
-    LogStartDateSettings,
 };
+
+pub mod loaded_logs;
 
 pub fn show_settings_grid(
     ui: &mut egui::Ui,
@@ -16,8 +17,7 @@ pub fn show_settings_grid(
     line_width: &mut f32,
     axis_cfg: &mut AxisConfig,
     plot_visibility_cfg: &mut PlotVisibilityConfig,
-    log_start_date_settings: &mut [LogStartDateSettings],
-    show_loaded_logs: &mut bool,
+    mut loaded_logs_ui: LoadedLogsUi<'_>,
 ) {
     egui::Grid::new("settings").show(ui, |ui| {
         ui.label("Line width");
@@ -54,98 +54,5 @@ pub fn show_settings_grid(
 
         ui.end_row();
     });
-    if !log_start_date_settings.is_empty() {
-        let show_loaded_logs_text = format!(
-            "{} Loaded logs",
-            if *show_loaded_logs {
-                regular::EYE
-            } else {
-                regular::EYE_SLASH
-            }
-        );
-        ui.toggle_value(show_loaded_logs, show_loaded_logs_text);
-
-        if *show_loaded_logs {
-            egui::Grid::new("loaded_logs").show(ui, |ui| {
-                for (i, settings) in log_start_date_settings.iter_mut().enumerate() {
-                    if (i + 1) % 6 == 0 {
-                        ui.end_row();
-                    }
-                    log_date_settings_ui(ui, settings);
-                }
-            });
-        }
-    }
-}
-
-pub fn log_date_settings_ui(ui: &mut egui::Ui, settings: &mut LogStartDateSettings) {
-    let log_name_date = format!("{} [{}]", settings.log_id, settings.start_date.date_naive());
-    let button_resp = ui.button(log_name_date.clone());
-    if button_resp.clicked() {
-        settings.clicked = !settings.clicked;
-    }
-    if button_resp.hovered() {
-        button_resp.on_hover_text("Click to modify log settings");
-    }
-
-    if settings.tmp_date_buf.is_empty() {
-        settings.tmp_date_buf = settings
-            .start_date
-            .format("%Y-%m-%d %H:%M:%S%.f")
-            .to_string();
-    }
-    if settings.clicked {
-        log_settings_window(ui, settings, &log_name_date);
-    }
-}
-
-fn log_settings_window(ui: &egui::Ui, settings: &mut LogStartDateSettings, log_name_date: &str) {
-    // State of window bound to the 'X'-button that closes the window
-    let mut open = true;
-    egui::Window::new(RichText::new(log_name_date).size(20.0).strong())
-        .collapsible(false)
-        .movable(false)
-        .open(&mut open)
-        .anchor(egui::Align2::CENTER_CENTER, egui::Vec2::ZERO)
-        .show(ui.ctx(), |ui| {
-            ui.vertical_centered(|ui| {
-                ui.label("Modify the start date to offset the plots of this log");
-                ui.label(format!("original date: {}", settings.original_start_date));
-                ui.label(RichText::new("YYYY-mm-dd HH:MM:SS.ms").strong());
-                let response = ui.add(TextEdit::singleline(&mut settings.tmp_date_buf));
-                if response.changed() {
-                    log::debug!("Changed to {}", settings.tmp_date_buf);
-                    match NaiveDateTime::parse_from_str(
-                        &settings.tmp_date_buf,
-                        "%Y-%m-%d %H:%M:%S%.f",
-                    ) {
-                        Ok(new_dt) => {
-                            settings.err_msg.clear();
-                            settings.new_date_candidate = Some(new_dt);
-                        }
-                        Err(e) => {
-                            settings.err_msg = format!("⚠ {e} ⚠");
-                        }
-                    };
-                }
-                if settings.err_msg.is_empty() {
-                    if let Some(new_date) = settings.new_date_candidate {
-                        if ui.button("Apply").clicked() || ui.input(|i| i.key_pressed(Key::Enter)) {
-                            settings.start_date = new_date.and_utc();
-                            settings.date_changed = true;
-                            log::info!("New date: {}", settings.start_date);
-                        }
-                    }
-                } else {
-                    ui.label(settings.err_msg.clone());
-                }
-                if ui.button("Cancel").clicked() {
-                    settings.clicked = false;
-                }
-            })
-        });
-
-    if !open || ui.ctx().input(|i| i.key_pressed(Key::Escape)) {
-        settings.clicked = false;
-    }
+    loaded_logs_ui.show(ui);
 }
